Finally got the build working on MSVC - hopefully it still works on non-windows platforms

+++ main/Wall.hpp
@@ -6,7 +6,7 @@ class Wall : public WorldObject
 {
 public:
 	Wall();
-	Wall(Point location, size_t width, size_t height);
+	Wall(Point location, unsigned long width, unsigned long height);
 
 	void CollisionHandler(const WorldObject&);
 };
